apps/CORE_UI/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
starlette
sse-starlette
httpx
requests
orjson
ujson
SQLAlchemy
aiosqlite
alembic
cachetools
python-dateutil
typing-extensions
tenacity
Jinja2
python-dotenv
openai>=1.0.0
faker
pyjwt
passlib[bcrypt]
bcrypt
email-validator
fastapi-cors
